changing float to int

+++ Prog03.py
@@ -1,8 +1,8 @@
 #Create a program that ask user to input 2 numbers. Print the sum of the two numbers.
 
 #input 2 numbers
-num_1 = float(input("Enter 1st Number: "))
-num_2 = float(input("Enter 2nd Number: "))
+num_1 = int(input("Enter 1st Number: "))
+num_2 = int(input("Enter 2nd Number: "))
 
 #Print the sum of the two numbers
 sum = num_1 + num_2
