+++ Prog04.py
@@ -1,8 +1,8 @@
 #Create a program that ask user to input 2 numbers. Print the product of the two numbers.
 
 #input 2 numbers
-num_1 = float(input("Enter 1st Number: "))
-num_2 = float(input("Enter 2nd Number: "))
+num_1 = int(input("Enter 1st Number: "))
+num_2 = int(input("Enter 2nd Number: "))
 
 #Print the product of the two numbers
 product = num_1 * num_2
